Add WAV reader and writer functionality
- Implemented wav_reader.c and wav_reader.h for parsing WAV, RF64, and BW64 headers, extracting metadata, and reading audio samples.
- Added wav_types.h to define data structures for WAV file metadata.
- Created wav_writer.c and wav_writer.h for writing standard and RF64 WAV headers, including support for extensible formats and fact chunks.
- Introduced functions for reading channel samples and updating headers with actual data sizes.

+++ src/Config/filter_config_types.h
@@ -200,8 +200,8 @@ typedef struct {
   char format[32];     /**< Sample format (for raw files). */
   int channel;         /**< Channel to use from multi-channel files. */
   int length;          /**< Max number of samples to read. */
-  int skip_bytes_lines; /**< Bytes/lines to skip at start of file. */
-  int read_bytes_lines; /**< Bytes/lines to read from file. */
+  size_t skip_bytes_lines; /**< Bytes/lines to skip at start of file. */
+  size_t read_bytes_lines; /**< Bytes/lines to read from file. */
 } convolution_config_t;
 
 /**
@@ -333,7 +333,7 @@ typedef struct {
  * @brief Parameters for a Clipper filter.
  */
 typedef struct {
-  double clip_limit; /**< Clip limit (linear scale). */
+  double clip_limit; /**< Clip limit (dB scale). */
   bool soft_clip;    /**< Enable soft clipping. */
 } clipper_config_t;
